requirements.txt
torch
torchvision
numpy
opencv-python
tqdm
Pillow
faiss-cpu
ftfy
regex
git+https://github.com/openai/CLIP.git
